requirements.txt
fastapi
uvicorn
jinja2
pydantic
pypdf
mss
pyyaml
pytest
flake8
